telegram_bot/requirements.txt

python-telegram-bot==13.15
python-dotenv==1.0.0
aiohttp==3.9.1
